Split MainWindow up into several classes in a package
