+++ src/tk_mvc/__init__.py
@@ -1,4 +1,5 @@
 from tk_mvc.view import View
 from tk_mvc.window import BaseWindow
+from tk_mvc.window_part import WindowPart
 from tk_mvc.controller import BaseController
 from tk_mvc.event import Event
